requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
python-dotenv==1.0.0
email-validator==2.1.0
APScheduler==3.10.4
gunicorn
reportlab==4.0.7
psycopg2-binary==2.9.9